pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xrdpattern"
version = "0.9.7"
description = "Python library for XrdPatterns including file import, file export and postprocessing functionalities"
readme = "README.md"
authors = [{ name = "Daniel Hollarek", email = "daniel.hollarek@googlemail.com" }]
license = { file = "LICENSE.txt" }
urls = { Homepage = "https://github.com/aimat-lab/xrdpattern" }
dependencies = [
    "numpy < 2.0.0",
    "scipy",
    "matplotlib",
    "pymatgen",
    "torch",
    "gemmi",
    "pandas",
    "tensordict",
    "orjson",
    "openpyxl",
    "pathvalidate",
    "progressbar2"
]


[tool.deptry.per_rule_ignores]
DEP002 = ["openpyxl"]
DEP001 = ["mpl_toolkits", "xylib"]

[tool.setuptools]
ext-modules = [{name = "_xylib",sources = ["xylib/philips_udf.cpp","xylib/uxd.cpp","xylib/winspec_spe.cpp","xylib/xrdml.cpp","xylib/xfit_xdd.cpp","xylib/cache.cpp","xylib/bruker_raw.cpp","xylib/spectra.cpp","xylib/riet7.cpp","xylib/csv.cpp","xylib/specsxy.cpp","xylib/bruker_spc.cpp","xylib/pdcif.cpp","xylib/canberra_mca.cpp","xylib/xylib.cpp","xylib/philips_raw.cpp","xylib/vamas.cpp","xylib/rigaku_dat.cpp","xylib/util.cpp","xylib/cpi.cpp","xylib/text.cpp","xylib/dbws.cpp","xylib/canberra_cnf.cpp","xylib/xsyg.cpp","xylib/chiplot.cpp", "xylib.i"],include-dirs = ["."],language = "c++",swig-opts = ["-c++"]}]
py-modules = ["xylib"]


[tool.setuptools.packages.find]
include = ["xrdpattern*"]


[tool.setuptools.package-data]
"xrdpattern.crystal.atomic_constants" = ["*"]
"xrdpattern.crystal.cifs" = ["*"]
"xrdpattern.parsing.examples" = ["*"]
"xrdpattern.tools" = ["*"]
